Implementing requested changes
